+++ DevProject/ProjectQ_Server/WebServer/WebServer/WebServer/Entity/Shop.py
@@ -6,23 +6,52 @@ Created on 2018. 5. 13.
 from Entity.Define import SHOP, ITEM_CONTANIER
 from MetaDataMgr import metaData
 from Route import Common
+from Entity import Define
 import Route.Common
+import DB
 
 class ShopBase(object):
     def __init__(self):
         self.itemMetas = metaData.getMetaDatas("ShopItem")
         
-    
     def getMetaData(self, itemId):
         return self.itemMetas[itemId]
     
-    def buyProduct(self, buyProductId, userObject):
-        
-        itemContainer = userObject.getData(ITEM_CONTANIER)
-        
+    def buyProduct(self, buyProductId, buyProductCount, userObject):
+
         if not buyProductId in self.itemMetas:
             return Common.respHandler.errorResponse(Route.Define.ERROR_NOT_FOUND_ITEM)
         
-        itemInfo = self.itemMetas[buyProductId]
+        buyItemInfo = self.itemMetas[buyProductId]
+        accountInfo = userObject.getData(Define.ACCOUNT_INFO)
         
-        findItem = itemContainer.getItemById(buyProductId)
+        priceValue = buyItemInfo['Price']
+        
+        if priceValue > accountInfo.gameMoney:
+            return Route.Define.ERROR_NOT_ENOUGH_MONEY
+        
+        itemContainer = userObject.getData(ITEM_CONTANIER)
+        
+        #1 == no stock item, 0 == stock item
+        type = buyItemInfo['ItemType']
+        if type == 1:
+            findItem = itemContainer.getItemById(buyProductId)
+            if not findItem is None: 
+                return Route.Define.ERROR_ALREADY_BUY_NO_STOCK_ITEM
+            
+        itemId = buyItemInfo['Index']
+        out_ItemIdx = 0
+        
+        #outputParams = Range(>=, <)
+        try:
+            resultDB = DB.dbConnection.executeStoredProcedure("Game_Item_BuyProduct", (itemId, accountInfo.accountId, buyProductCount, priceValue, out_ItemIdx), (4, 5))
+        except Exception as e:
+            print(str(e))
+            return Route.Define.ERROR_DB
+        
+        accountInfo.gameMoney -= priceValue
+        accountInfo.syncToResp()
+        itemContainer.setItem(resultDB[0], itemId, buyProductCount)
+        
+        return Route.Define.OK_SHOP_BUY_PRODUCT
+
