+++ DevProject/ProjectQ_Server/WebServer/WebServer/WebServer/Route/Define.py
@@ -10,6 +10,7 @@ ERROR_NOT_FOUND_SESSION = 103
 ERROR_INPUT_PARAMS = 102
 ERROR_INVALID_ACCESS = 104
 ERROR_OUT_OF_RANGE = 105
+ERROR_DB = 106
 
 #Login Create
 ERROR_CREATE_NOT_LOGIN = 1001
@@ -23,7 +24,9 @@ OK_SUCCESS = 2003
 #Shop
 ERROR_INVALID_BUY_PRODUCT = 3001
 ERROR_NOT_ENOUGH_MONEY = 3002
-ERROR_NOT_FOUND_ITEM =3003 
+ERROR_NOT_FOUND_ITEM = 3003
+OK_SHOP_BUY_PRODUCT = 3004
+ERROR_ALREADY_BUY_NO_STOCK_ITEM = 3005
 
 #Inventory
 ERROR_ALREADY_EQUIP_ITEM = 4001
