+++ DevProject/ProjectQ_Server/WebServer/WebServer/WebServer/Route/Shop.py
@@ -4,35 +4,51 @@ from flask import jsonify, request, session
 import DB
 import Util
 import Route.Common
-from Route import Common
 import Entity.User
-from Entity import serverCachedObject, userCachedObjects
 
-class TradeShop(Resource, Common.BaseRoute):
-    def get(self):
+from Entity.Define import ITEM_CONTANIER
+from Route import Common
+from Entity import serverCachedObject, userCachedObjects
+from Entity import Define
+
+class ShopBuyProduct(Resource, Common.BaseRoute):
+    def post(self):
         session = self.getSession(request)
         if session is None:
             return jsonify(Common.respHandler.errorResponse(Route.Define.ERROR_NOT_FOUND_SESSION))
-        
-    def put(self):
-        session = self.getSession(request)
-        if session is None:
-            return jsonify(Common.respHandler.errorResponse(Route.Define.ERROR_NOT_FOUND_SESSION))
-        
-        Route.parser.add_argument("buyId")
+
+        Route.parser.add_argument("buyProductId")
+        Route.parser.add_argument("buyProductCount")
         args = Route.parser.parse_args()
-        
-        if not "buyId" in args:
+
+        if not args["buyProductId"]:
             return jsonify(Common.respHandler.errorResponse(Route.Define.ERROR_INPUT_PARAMS))
         
+        if not args["buyProductCount"]:
+            return jsonify(Common.respHandler.errorResponse(Route.Define.ERROR_INPUT_PARAMS))
+
         if not session in userCachedObjects:
             return jsonify(Common.respHandler.errorResponse(Route.Define.ERROR_INVALID_ACCESS))
-        
+
         userObject = userCachedObjects[session]
-        
+
         shop = serverCachedObject[Entity.Define.SHOP]
-        resp = shop.buyProduct(int(args["buyId"]), userObject)
-        return jsonify(resp)
+        
+        buyId = int(args["buyProductId"])
+        buyProductCount = int(args["buyProductCount"])
+        
+        respCode = shop.buyProduct(buyId, buyProductCount, userObject)
+        
+        if not respCode is Route.Define.OK_SHOP_BUY_PRODUCT:
+            return jsonify(Common.respHandler.errorResponse(respCode))
+        
+        accountInfo = userObject.getData(Define.ACCOUNT_INFO)
+        itemContainer = userObject.getData(ITEM_CONTANIER)
+        
+        Common.respHandler.mergeResp(accountInfo.getResp())
+        Common.respHandler.mergeResp(itemContainer.getContainerResp())
+        
+        return Common.respHandler.getResponse(Route.Define.OK_SHOP_BUY_PRODUCT)
