+++ DevProject/ProjectQ_Server/WebServer/WebServer/WebServer/Route/__init__.py
@@ -1,7 +1,7 @@
 from Route import Login
 from flask_restful import reqparse
 from Route.Login import Login
-from Route.Shop import TradeShop
+from Route.Shop import ShopBuyProduct
 from Route.Mail import MailPostRead
 from Route.Mail import MailPostAccept
 from Route.Mail import MailWrite
@@ -10,7 +10,7 @@ from Route.User import InventoryUnEquip
 
 route_dict = { 
     Login:'/loginInfo', 
-    TradeShop:'/tradeShop',
+    ShopBuyProduct:'/shop/buyProduct',
     MailPostRead : '/mailPost/read',
     MailPostAccept : '/mailPost/accept',
     MailWrite : '/mailPost/write',
